Refactor solution
Refactor the solution to use ranges and memoize the distances to an end-point from a given node and location on the path.

+++ tests/AdventOfCode.Tests/Puzzles/Y2023/Day08Tests.cs
@@ -64,13 +64,13 @@ public sealed class Day08Tests(ITestOutputHelper outputHelper) : PuzzleTest(outp
 
     [Theory]
     [MemberData(nameof(TestCases))]
-    public void Y2023_Day08_WalkNetwork_Returns_Correct_Value(string[] network, bool asGhost, int expected)
+    public void Y2023_Day08_WalkNetwork_Returns_Correct_Value(string[] network, bool asGhost, long expected)
     {
         // Arrange
         using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(15));
 
         // Act
-        int actual = Day08.WalkNetwork(network, asGhost, cts.Token);
+        long actual = Day08.WalkNetwork(network, asGhost, cts.Token);
 
         // Assert
         actual.ShouldBe(expected);
